+++ docs/sphinx/changelog.rst
@@ -22,6 +22,7 @@ New Features
 - Added the balance of volume fraction (:pull:`17`). By `Nathan Miller`_.
 - Added material model version for the balance equations. (:pull:`18`). By `Nathan Miller`_.
 - Added a constraint equation enabling complex relationships between the material state and the internal energy. (:pull:`19`). By `Nathan Miller`_.
+- Added a constraint equation for the displacement. (:pull:`20`). by `Nathan Miller`_.
 
 Internal Changes
 ================